+++ src/containers/CategoriesContainer.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 
 import { useDispatch, useSelector } from 'react-redux';
 
-import { selectCategory, loadRestaurants } from './actions';
+import { selectCategory, loadRestaurants } from '../actions';
 
-import { get } from './utils';
+import { get } from '../utils';
 
 export default function CategoriesContainer() {
   const dispatch = useDispatch();
