+++ src/containers/RegionsContainer.jsx
@@ -2,9 +2,9 @@ import React from 'react';
 
 import { useDispatch, useSelector } from 'react-redux';
 
-import { selectRegion, loadRestaurants } from './actions';
+import { selectRegion, loadRestaurants } from '../actions';
 
-import { get } from './utils';
+import { get } from '../utils';
 
 export default function RegionsContainer() {
   const dispatch = useDispatch();
