+++ src/containers/RegionsContainer.test.jsx
@@ -5,7 +5,7 @@ import { render, fireEvent } from '@testing-library/react';
 import { useDispatch, useSelector } from 'react-redux';
 
 import RegionsContainer from './RegionsContainer';
-import REGIONS from '../fixtures/regions';
+import REGIONS from '../../fixtures/regions';
 
 describe('RegionsContainer', () => {
   const dispatch = jest.fn();
