+++ src/pages/RestaurantDetailPage.test.jsx
@@ -7,7 +7,7 @@ import { Router } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 
 import RestaurantDetailPage from './RestaurantDetailPage';
-import RESTAURANT_DETAIL from '../fixtures/restaurantDetail';
+import RESTAURANT_DETAIL from '../../fixtures/restaurantDetail';
 
 describe('RestaurantDetailPage', () => {
   const dispatch = jest.fn();
