+++ src/pages/RestaurantsPage.jsx
@@ -2,13 +2,11 @@ import React, { useEffect } from 'react';
 
 import { useDispatch } from 'react-redux';
 
-import RegionsContainer from './RegionsContainer';
-import CategoriesContainer from './CategoriesContainer';
-import RestaurantsContainer from './RestaurantsContainer';
+import RegionsContainer from '../containers/RegionsContainer';
+import CategoriesContainer from '../containers/CategoriesContainer';
+import RestaurantsContainer from '../containers/RestaurantsContainer';
 
-import {
-  loadInitialData,
-} from './actions';
+import { loadInitialData } from '../actions';
 
 export default function RestaurantsPage() {
   const dispatch = useDispatch();
